refactor: split database tests into separate files
- Split database-logic.spec.ts into 3 focused test files:
  - create-bookmark.spec.ts: Tests for bookmark creation logic
  - bookmark-validation.spec.ts: Tests for limits and validation
  - get-bookmark.spec.ts: Tests for bookmark retrieval
- Better organization and readability
- All 15 tests pass (4 + 5 + 6 tests respectively)
